Feat/membuat akun - separating page components (#13)
* [RED] Add API for Membuat Akun, add test for message dialog in Membuat Akun FE.

* [GREEN] Add dialog for error message.

* [RED] Add test for login page.

* [GREEN] Add Page for Login.

* [GREEN] Fix button testid not appearing.

* [RED] Add test for recover account page.

* [GREEN] Add page for recover email, fix recover email page test.

* [RED] Add test for Register page, add more test for login and recover email page.

* [GREEN] Add register page, fix login and recover account page.

* [RED] Create test for verify email page, fix register test naming.

* [GREEN] Implement verify email page.

* [REFACTOR] Fix Code Quality issue and Security Hotspot, delete unused dialog.

* [REFACTOR] Fix issues and security hotspot.

* [CHORES] Fix userSlice and authentication.ts to accomodate access token.

* [RED] Add new test and modify old test for Login Page.

* [GREEN] Finish implementing Login Page component by refactoring the old ones.

* [REFACTOR] Fix issue on AccountRecoveryRequestForm.

* [RED] Add test and modify test for Recover Account Page.

* [GREEN] Add AccountRecoveryForm and change account recovery page.

* [RED] Add test adn modify test for Email Verification Page.

* [GREEN] Add VerifyForm.tsx and change email verification page.

* [RED] Add test and modify test for register page

* [GREEN] Finish separating register page into multiple components and passed the test.
