+++ __tests__/register/register.test.tsx
@@ -5,6 +5,8 @@ import {
   useLazySendEmailVerficationQuery,
 } from '@/redux/api/authApi'
 import '@testing-library/jest-dom'
+import RegisterForm from '@/app/register/RegisterForm'
+import RegisterPageTitle from '@/app/register/RegisterPageTitle'
 
 jest.mock('@/redux/api/authApi', () => ({
   useRegisterMutation: jest.fn(),
@@ -21,15 +23,43 @@ describe('Test for register page', () => {
     ])
   })
 
-  it('renders register form', () => {
+  it('renders register page', () => {
     const { getByTestId } = render(<Register />)
+    expect(getByTestId('register-form')).toBeInTheDocument()
+    expect(getByTestId('register-title')).toBeInTheDocument()
+  })
+})
+
+describe('Test for register page title', () => {
+  it('renders register page title', () => {
+    const { getByTestId } = render(<RegisterPageTitle />)
+    expect(getByTestId('register-title')).toBeInTheDocument()
+  })
+})
+
+describe('Test for register form', () => {
+  beforeEach(() => {
+    const mockRegister = jest.fn().mockResolvedValue({ data: {} })
+    ;(useRegisterMutation as jest.Mock).mockReturnValue([mockRegister])
+    const mockSendEmailVerification = jest.fn().mockResolvedValue({ data: {} })
+    ;(useLazySendEmailVerficationQuery as jest.Mock).mockReturnValue([
+      mockSendEmailVerification,
+    ])
+  })
+
+  it('renders register form', () => {
+    const { getByTestId } = render(<RegisterForm />)
     expect(getByTestId('register-form')).toBeInTheDocument()
   })
 
-  it('submits register form successfully', async () => {
+  it('submits register form successfully and calls send email verification API', async () => {
     const mockRegister = jest.fn().mockResolvedValue({ data: {} })
     ;(useRegisterMutation as jest.Mock).mockReturnValue([mockRegister])
-    const { getByTestId } = render(<Register />)
+    const mockSendEmailVerification = jest.fn().mockResolvedValue({ data: {} })
+    ;(useLazySendEmailVerficationQuery as jest.Mock).mockReturnValue([
+      mockSendEmailVerification,
+    ])
+    const { getByTestId } = render(<RegisterForm />)
     fireEvent.change(getByTestId('email-input'), {
       target: { value: 'email@email.com' },
     })
@@ -41,6 +71,9 @@ describe('Test for register page', () => {
     })
     fireEvent.submit(getByTestId('register-form'))
     await waitFor(() => expect(mockRegister).toHaveBeenCalledTimes(1))
+    await waitFor(() =>
+      expect(mockSendEmailVerification).toHaveBeenCalledTimes(1)
+    )
   })
 
   it('shows error message when register account fails due to taken username and/or email', async () => {
@@ -49,7 +82,7 @@ describe('Test for register page', () => {
       .fn()
       .mockResolvedValue({ error: { data: { msg: errorMessage } } })
     ;(useRegisterMutation as jest.Mock).mockReturnValue([mockRegister])
-    const { getByTestId, getByText } = render(<Register />)
+    const { getByTestId, getByText } = render(<RegisterForm />)
     fireEvent.change(getByTestId('email-input'), {
       target: { value: 'email@email.com' },
     })
@@ -68,7 +101,7 @@ describe('Test for register page', () => {
     const errorMessage = 'Unknown Error!'
     const mockRegister = jest.fn().mockResolvedValue({ error: {} })
     ;(useRegisterMutation as jest.Mock).mockReturnValue([mockRegister])
-    const { getByTestId, getByText } = render(<Register />)
+    const { getByTestId, getByText } = render(<RegisterForm />)
     fireEvent.change(getByTestId('email-input'), {
       target: { value: 'email@email.com' },
     })
