+++ src/app/login/recover-account/page.tsx
@@ -1,67 +1,8 @@
 'use client'
-import { useState } from 'react'
-import {
-  Box,
-  Typography,
-  Dialog,
-  DialogTitle,
-  DialogActions,
-  DialogContent,
-} from '@mui/material'
-import { SubmitHandler, useForm } from 'react-hook-form'
-import { useSendChangePasswordMutation } from '@/redux/api/authApi'
-import { Input } from '@/components/elements/Forms/input'
-import Link from 'next/link'
-
-type ChangePassFormType = {
-  token: string
-  new_password: string
-  email: string
-}
+import { Box } from '@mui/material'
+import AccountRecoveryForm from './AccountRecoveryForm'
 
 export default function RecoverAccount() {
-  const defaultValues: ChangePassFormType = {
-    token: '',
-    new_password: '',
-    email: '',
-  }
-
-  const [openPrompt, setOpenPrompt] = useState(false)
-  const [message, setMessage] = useState('')
-  const [openPopup, setOpenPopup] = useState(false)
-  const [errorMessage, setErrorMessage] = useState('')
-  const [dialogTitle, setDialogTitle] = useState('')
-
-  const [changePass] = useSendChangePasswordMutation()
-
-  const methods = useForm<ChangePassFormType>({ defaultValues: defaultValues })
-  const { control, handleSubmit } = methods
-
-  const onSubmit: SubmitHandler<ChangePassFormType> = async (data) => {
-    await changePass({ ...data }).then((res) => {
-      if (res) {
-        if ('data' in res) {
-          setOpenPrompt(true)
-          setMessage('Success! Please Log In with your new password.')
-          setDialogTitle('Account Recovery')
-        } else if ('data' in res.error) {
-          const errorData = res.error.data as { msg: string }
-          setErrorMessage(errorData.msg)
-          setOpenPopup(true)
-          setDialogTitle('Error')
-        } else {
-          setErrorMessage('Unknown Error!')
-          setOpenPopup(true)
-          setDialogTitle('Error')
-        }
-      }
-    })
-  }
-
-  const handleClosePopUP = () => {
-    setOpenPopup(false)
-  }
-
   return (
     <Box
       data-testid="recover"
@@ -76,71 +17,7 @@ export default function RecoverAccount() {
       justifyContent={'center'}
       padding={{ xs: '4em 4em', lg: '4em 4em' }}
     >
-      <form
-        className="flex flex-col gap-3"
-        onSubmit={handleSubmit(onSubmit)}
-        style={{ flex: 1 }}
-        data-testid="recover-form"
-      >
-        <Typography fontWeight={'bold'} variant="h3" align="center">
-          Account Recovery
-        </Typography>
-        <Typography fontWeight={'bold'} align="center">
-          Please Enter Your Email, Token Sent to Your Email, and Your New
-          Password
-        </Typography>
-        <Input
-          control={control}
-          name="email"
-          placeholder="Enter Email"
-          required
-          data-testid="email-input"
-        />
-        <Input
-          control={control}
-          name="token"
-          placeholder="Enter token"
-          required
-          data-testid="token-input"
-        />
-        <Input
-          type="password"
-          control={control}
-          name="new_password"
-          placeholder="Enter new password"
-          required
-          data-testid="password-input"
-        />
-        <button type="submit" className="bg-teal-50">
-          Recover Account
-        </button>
-      </form>
-
-      <Dialog open={openPrompt} data-testid="recover-login-redirect-msg">
-        <DialogTitle>{dialogTitle}</DialogTitle>
-        <DialogContent>
-          <p>{message}</p>
-        </DialogContent>
-        <DialogActions>
-          <Link href={'/login'}>Continue</Link>
-        </DialogActions>
-      </Dialog>
-
-      <Dialog
-        open={openPopup}
-        onClose={handleClosePopUP}
-        data-testid="recover-dialog-error-msg"
-      >
-        <DialogTitle>{dialogTitle}</DialogTitle>
-        <DialogContent>
-          <p>{errorMessage}</p>
-        </DialogContent>
-        <DialogActions>
-          <button data-testid="button-error" onClick={handleClosePopUP}>
-            Close
-          </button>
-        </DialogActions>
-      </Dialog>
+      <AccountRecoveryForm data-testid="recover-form" />
     </Box>
   )
 }
